Trouble

+++ lib/Effects/Effects/Strobe.h
@@ -2,7 +2,7 @@
 
 void strobeEffectLoop(struct EffectDataPacket *effectState, CRGB colorOff, CRGB colorOn, uint16_t loopMs)
 {
-  const uint8_t loopPosition = effectState->loopPosition % loopMs;
+  const uint16_t loopPosition = effectState->loopPosition % loopMs;
 
   for (uint16_t i = 0; i < LED_CNT; i++)
   {
